+++ example/lib/main.dart
@@ -93,31 +93,67 @@ class _MyHomePageState extends State<MyHomePage> {
               ),
               CubitFormTextField(
                 formFieldCubit: formReal.login,
-                hintText: 'name',
-                labeText: 'login',
+                decoration: InputDecoration(
+                  hintText: 'name',
+                  labelText: 'login',
+                ),
               ),
               CubitFormTextField(
                 formFieldCubit: formReal.password,
                 obscureText: true,
-                hintText: 'another text',
-                labeText: 'password',
+                decoration: InputDecoration(
+                  hintText: 'another text',
+                  labelText: 'password',
+                ),
               ),
               BlocBuilder<Real, FormCubitState>(
                   cubit: formReal,
                   builder: (context, state) {
-                    return Center(
-                      child: RaisedButton(
-                        child: Text('submit'),
-                        onPressed: state.isSubmitting
-                            ? null
-                            : () {
-                                formReal.trySubmit();
-                              },
-                      ),
-                    );
+                    print(state.isErrorShown);
+
+                    if (state.isFormValid || !state.isErrorShown) {
+                      return Center(
+                        child: RaisedButton(
+                          child:
+                              Text(state.isSubmitting ? 'Submiting' : 'Submit'),
+                          onPressed:
+                              state.isSubmitting ? null : formReal.trySubmit,
+                        ),
+                      );
+                    } else {
+                      return Center(
+                        child: _Error(
+                          child: Text('Not valid'),
+                        ),
+                      );
+                    }
                   })
             ],
           ),
+        ),
+      ),
+    );
+  }
+}
+
+class _Error extends StatelessWidget {
+  const _Error({Key key, this.child}) : super(key: key);
+
+  final Widget child;
+  @override
+  Widget build(BuildContext context) {
+    return Material(
+      elevation: 2,
+      child: Container(
+        constraints: BoxConstraints(minWidth: 88.0, minHeight: 36.0),
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(3),
+          color: Colors.red,
+        ),
+        child: Center(
+          widthFactor: 1.0,
+          heightFactor: 1.0,
+          child: child,
         ),
       ),
     );