+++ lib/cubit_form.dart
@@ -4,7 +4,8 @@ export 'src/components/cubit_form_field.dart';
 export 'src/components/cubit_form_text_field.dart';
 
 export 'src/logic/field_cubit/field_cubit.dart';
-export 'src/logic/models/validadation.dart';
+export 'src/logic/models/validation.dart';
+export 'src/logic/models/validations.dart';
 
 export 'src/logic/form_cubit/form_cubit.dart';
 export 'package:flutter_bloc/flutter_bloc.dart';