+++ lib/src/logic/form_cubit/form_cubit_state.dart
@@ -29,5 +29,6 @@ class FormCubitState extends Equatable {
   }
 
   @override
-  List<Object> get props => [isErrorShown, isFormValid];
+  List<Object> get props =>
+      [isErrorShown, isFormValid, isSubmitting, isSubmitted];
 }