+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.0.6
+  - fix
+
+## 0.0.5
+  - fix
+
+## 0.0.4
+  - fix
+
+## 0.0.3
+  - fix
+
+## 0.0.2
+
+- add validations:
+    - string: max, min, fixed length,
+    - email validation
+    - not null validation
+    - required: int, string
+
+- fix field cubit;
+
 ## 0.0.1
 
 - Initial version